Extract ferment event wiring
