+++ src/extensions/ferment/events.ts
@@ -1,0 +1,157 @@
+import type { ExtensionAPI } from "@earendil-works/pi-coding-agent"
+import { clearFermentCache } from "../../ferment/store.js"
+import { extractContextualOptions, extractTrailingQuestion } from "./contextual-options.js"
+import { buildPlannerSupplement } from "./planner-supplement.js"
+import { resumeFerment } from "./resume.js"
+import { defaultFermentRuntime } from "./runtime.js"
+import { confirmPendingScope } from "./scoping-confirmation.js"
+import { clearAllPendingScopes } from "./scoping.js"
+import {
+	captureJudgeContext,
+	clearAllScopingGates,
+	clearAllStepStarts,
+	getActive,
+	isRestoringModel,
+	markHumanInput,
+	setActive,
+	setRestoringModel,
+} from "./state.js"
+import { applyAndPersist } from "./tool-helpers.js"
+
+type AssistantContentPart = { type: string; text?: string; name?: string }
+
+function hasToolCall(content: AssistantContentPart[], toolName: string): boolean {
+	return content.some((c) => c.type === "toolCall" && c.name === toolName)
+}
+
+function extractPromptTextAfterLastToolCall(content: AssistantContentPart[]): string {
+	const lastToolCall = content.findLastIndex((c) => c.type === "toolCall")
+	return content
+		.slice(lastToolCall + 1)
+		.filter((c) => c.type === "text")
+		.map((c) => c.text ?? "")
+		.join("")
+		.trimEnd()
+}
+
+export function registerFermentEvents(pi: ExtensionAPI): void {
+	pi.on("session_start", async (_event, ctx) => {
+		if (process.env.KIMCHI_SUBAGENT === "1") return
+		clearAllStepStarts()
+		clearAllScopingGates()
+		clearAllPendingScopes()
+		clearFermentCache()
+
+		const envId = process.env.KIMCHI_ACTIVE_FERMENT
+		if (envId) {
+			resumeFerment(pi, envId, ctx)
+		} else {
+			setActive(undefined)
+		}
+	})
+
+	pi.on("session_shutdown", async () => {
+		if (process.env.KIMCHI_SUBAGENT === "1") return
+		const f = getActive()
+		if (!f) return
+		if (f.status === "running" || f.status === "planned") {
+			applyAndPersist(f.id, { type: "pause" })
+		}
+	})
+
+	pi.on("input", async (event) => {
+		if (event.source === "interactive") {
+			markHumanInput()
+		}
+	})
+
+	pi.on("before_agent_start", async (event) => {
+		const f = getActive()
+		if (!f) return {}
+		if (f.status === "paused") {
+			const pausedSupplement = `\n\n## Ferment Paused\n\nFerment "${f.name}" is paused by the user. Do NOT call any ferment tools (activate_phase, start_step, complete_step, etc.) — they will be rejected. Acknowledge any pending question briefly and wait for the user to resume with /auto.`
+			return { systemPrompt: `${event.systemPrompt}${pausedSupplement}` }
+		}
+		if (f.status !== "running") return {}
+		const supplement = buildPlannerSupplement()
+		return { systemPrompt: `${event.systemPrompt}${supplement}` }
+	})
+
+	pi.on("model_select", async (event, ctx) => {
+		captureJudgeContext(ctx?.model, ctx?.modelRegistry)
+		const f = getActive()
+		if (!f || f.status !== "running") return
+		if (isRestoringModel()) return
+
+		if (event.previousModel) {
+			setRestoringModel(true)
+			pi.setModel(event.previousModel)
+				.catch(() => {})
+				.finally(() => {
+					setRestoringModel(false)
+				})
+		}
+		ctx.ui.notify(
+			`Model switching is locked while ferment "${f.name}" is running. Finish or abandon the ferment first.`,
+			"warning",
+		)
+	})
+
+	pi.on("turn_end", async (event, ctx) => {
+		captureJudgeContext(ctx?.model, ctx?.modelRegistry)
+		const f = getActive()
+		if (!f) return
+		if (f.mode === "exec") return
+		if (f.status !== "draft" && f.status !== "running") return
+		if (!ctx?.ui?.select || !ctx?.ui?.input) return
+		if (event.message.role !== "assistant") return
+
+		if (hasToolCall(event.message.content, "propose_phases")) return
+		const text = extractPromptTextAfterLastToolCall(event.message.content)
+		if (!text) return
+
+		const isDraft = f.status === "draft"
+		const yesLabel = isDraft ? "Yes, this looks right" : "Yes, proceed"
+		const noLabel = isDraft ? "No, revise" : "No, pause"
+
+		const title = extractTrailingQuestion(text)
+		const contextualOptions = extractContextualOptions(text)
+		if (!text.endsWith("?") && !contextualOptions) return
+		const options = contextualOptions
+			? [...contextualOptions, "Let me say something else"]
+			: [yesLabel, noLabel, "Let me say something else"]
+		const choice = await ctx.ui.select(title, options)
+		if (!choice) return
+
+		let reply: string
+
+		if (choice === "Let me say something else") {
+			const custom = ctx.ui.input ? await ctx.ui.input("Your message:", "") : undefined
+			if (!custom) return
+			reply = custom
+		} else if (choice === noLabel) {
+			reply = isDraft ? "No — please revise." : "No, pause for now."
+		} else if (contextualOptions?.includes(choice)) {
+			reply = choice
+		} else if (isDraft && choice === yesLabel) {
+			const outcome = confirmPendingScope(defaultFermentRuntime, f.id, undefined, "turn_end", f.name)
+			if (outcome.ok) {
+				ctx.ui.notify(
+					`Plan saved for "${outcome.outcome.ferment.name}". ${outcome.outcome.ferment.phases.length} phase(s) ready.`,
+				)
+				reply = `Plan saved by user confirmation — ${outcome.outcome.ferment.phases.length} phase(s) now in "planned" status. You can proceed with activate_phase when the user is ready, or wait for further instructions.`
+			} else if (outcome.error.code !== "MISSING_PENDING_PHASES" && outcome.error.code !== "MISSING_PENDING_SCOPE") {
+				ctx.ui.notify(`Failed to save plan: ${outcome.error.message}`)
+				reply = `Plan save failed: ${outcome.error.message}. Investigate the ferment state and try again.`
+			} else {
+				reply =
+					"User confirmed the plan but you never called propose_phases — there's nothing structured for the host to save. Call propose_phases now with the same plan you just showed, then end with 'Does this plan look right?' so the user can confirm again."
+			}
+		} else {
+			reply = "Yes, proceed."
+		}
+
+		markHumanInput()
+		void pi.sendUserMessage(reply, { deliverAs: "followUp" })
+	})
+}
